requirements.txt
python-dotenv
telethon
google-cloud-secret-manager
google-cloud-storage